Conflicts:
	src/Jitter_Optimize.cpp

+++ src/Jitter_Optimize.cpp
@@ -765,9 +765,9 @@ bool CJitter::FoldConstantOperation(STATEMENT& statement)
 	{
 		if (src1cst && src2cst)
 		{
-			int32 quotient = static_cast<int32>(src1cst->m_valueLow) / static_cast<int32>(src2cst->m_valueLow);
-			uint32 remainder = src1cst->m_valueLow % src2cst->m_valueLow;
-			uint64 result = static_cast<int64>(quotient) | (static_cast<uint64>(remainder) << 32);
+			uint32 quotient = static_cast<int32>(src1cst->m_valueLow) / static_cast<int32>(src2cst->m_valueLow);
+			uint32 remainder = static_cast<int32>(src1cst->m_valueLow) % static_cast<int32>(src2cst->m_valueLow);
+			uint64 result = static_cast<uint64>(quotient) | (static_cast<uint64>(remainder) << 32);
 			statement.op = OP_MOV;
 			statement.src1 = MakeSymbolRef(MakeConstant64(result));
 			statement.src2.reset();
